added cast for strtol used in atoi
strtol returns long, which must be cast to int for returning it from atoi

+++ src/stdlib/atoi.c
@@ -23,5 +23,5 @@
 #include <macros/NULL.h>
 
 int atoi(const char* nptr){
-    return strtol(nptr, NULL, 10);
+    return (int)strtol(nptr, NULL, 10);
 }
